dom/payments/test/test_bug1443914.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1443914
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1443914</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript">

  "use strict";
  SimpleTest.waitForExplicitFinish();

  var gUrl = SimpleTest.getTestFileURL('Bug1443914ChromeScript.js');
  var gScript = SpecialPowers.loadChromeScript(gUrl);

  function testFailHandler(message) {
    ok(false, message);
  }
  function testPassHandler(message) {
    ok(true, message);
  }
  gScript.addMessageListener("test-fail", testFailHandler);
  gScript.addMessageListener("test-pass", testPassHandler);

  // testing data declation
  const defaultMethods = [{
    supportedMethods: "basic-card",
    data: {
      supportedNetworks: ['unionpay', 'visa', 'mastercard', 'amex', 'discover',
                          'diners', 'jcb', 'mir',
      ],
      supportedTypes: ['prepaid', 'debit', 'credit'],
    },
  }, {
    supportedMethods: "testing-payment-method",
  }];

  const defaultDetails = {
    id: "test payment",
    total: {
      label: "Total",
      amount: {
        currency: "USD",
        value: "1.00"
      }
    },
    shippingOptions: [
      {
        id: "NormalShipping",
        label: "NormalShipping",
        amount: {
          currency: "USD",
          value: "10.00"
        },
        selected: false,
      },
    ],
  };

  const defaultOptions = {
    requestPayerName: true,
    requestPayerEmail: false,
    reqeustPayerPhone: false,
    requestShipping: true,
    shippingType: "shipping"
  };

  const updateDetails = {
    total: {
      label: "Total",
      amount: {
        currency: "USD",
        value: "1.00"
      }
    },
    error: "",
  };
  // testing function main body
  function test_bug1443914() {
    gScript.sendAsyncMessage("set-checking-shipping-options-ui-service");
    return new Promise((resolve, reject) => {
      const request = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      request.addEventListener("shippingaddresschange", (event) => {
        event.updateWith(updateDetails);
      });
      request.addEventListener("shippingoptionchange", (event) => {
        event.updateWith(updateDetails);
      });
      request.show().then(response => {
        response.complete("success").then(() => {
	  resolve();
	}).catch(error => {
	  ok(false, "Unexpected error: " + error.name);
	  resolve();
	});
      }).catch(error => {
        ok(false, "Unexpected error: " + error.name);
	resolve();
      }).finally(handler.destruct);
    });
  }

  function teardown() {
    gScript.addMessageListener("teardown-complete", function teardownCompleteHandler() {
      gScript.removeMessageListener("teardown-complete", teardownCompleteHandler);
      gScript.removeMessageListener("test-fail", testFailHandler);
      gScript.removeMessageListener("test-pass", testPassHandler);
      gScript.destroy();
      SimpleTest.finish();
    });
    gScript.sendAsyncMessage("teardown");
  }

  function runTests() {
    test_bug1443914()
    .then(teardown)
    .catch( e => {
      ok(false, "Unexpected error: " + e.name);
      SimpleTest.finish();
    });
  }

  window.addEventListener('load', function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
      ]
    }, runTests);
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1443914">Mozilla Bug 1443914</a>
</body>
</html>
